Refactor CustomExtension into top-level tests module
